requirements.txt
asyncio==3.4.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0